1、用户播放记录 2、视频主表和剧集

+++ KdyWeb.NetCore/KdyWeb.Entity/KdyUser/KdyUser.cs
@@ -6,7 +6,7 @@ namespace KdyWeb.Entity
     /// <summary>
     /// 注册用户
     /// </summary>
-    public class KdyUser : BaseEntity<int>
+    public class KdyUser : BaseEntity<long>
     {
         #region 常量
         /// <summary>
